First draft of ViewModel refactor
