+++ source/CoordinateConversion/CoordinateConversionLibrary/Helpers/Constants.cs
@@ -29,5 +29,7 @@ namespace CoordinateConversionLibrary
         public const string BroadcastCoordinateValues = "BROADCAST_COORDINATE_VALUES";
         public const string RequestOutputUpdate = "REQUEST_OUTPUT_UPDATE";
         public const string SetListBoxItemAddInPoint = "SET_LISTBOX_ITEM_ADDINPOINT";
+        public const string NewMapPointSelection = "NEW_MAP_POINT_SELECTION";
+        public const string SetCoordinateGetter = "SET_COORDINATE_GETTER";
     }
 }
