+++ source/CoordinateConversion/CoordinateConversionLibrary/Views/BatchCoordinateView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CoordinateConversionLibrary.Views.BatchCoordinateView"
+﻿<UserControl x:Class="CoordinateConversionLibrary.Views.CCCollectTabView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -12,8 +12,6 @@
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="/CoordinateConversionLibrary;component/MAResourceDictionary.xaml" />
             </ResourceDictionary.MergedDictionaries>
-            <!--<viewModels:MainViewModel x:Key="viewModelMain" />
-            <viewModels:BoolToOppositeBoolConverter x:Key="negateBool" />-->
             <Style BasedOn="{StaticResource {x:Type ListBoxItem}}" TargetType="{x:Type ListBoxItem}">
                 <Setter Property="IsSelected" Value="{Binding IsSelected, Mode=TwoWay}" />
             </Style>
