+++ source/CoordinateConversion/CoordinateConversionLibrary/Views/CoordinateConversionView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="CoordinateConversionLibrary.Views.CoordinateConversionView"
+﻿<UserControl x:Class="CoordinateConversionLibrary.Views.CCConvertTabView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
